Make a language def struct. Decode tells you state.
The language ought to be pluggable: Whether you want to use
one of the different families.  I think there are some (eg, cyrillic)
that use more states, so this type might need to be refactored into
something like a hashmap of state keys and charsets.

Return the state from Decode so it can be chained with incoming
chunks of data.  This probably should be an impl on a struct
that stores this state and the language in use.